Newside (#35)
* cash register

* more md fixes

Accidentally deleted the part from lessons.admin that had the autofill for slugs, reimplemented here. Also updated requirements.txt since it's been stale for a while now.

* md and static

* new css

* sidebar basic

Got the groundwork for the 2 layered sidebar thanks to this stack overflow post. Still need to do css and figure out convention for slugging it all.
https://stackoverflow.com/questions/38288280/iterate-through-an-object-in-django-template

* better style